+++ internal/storage/postgresql/migrations/168_add_changelog_v4_13_5.down.sql
@@ -1,0 +1,1 @@
+DELETE FROM changelogs WHERE version = '4.13.5';